// src/models/person.ts
export default interface Person {
    id: number;
    name: string;
    birth_year: number;
    gender: string;
    locked: boolean;
    language: string;
    photo: string;
    small_thumbnail: string;
    large_thumbnail: string;
    email: string;
    telephone_number: string;
    skype_name: string;
    website: string;
    facebook: string;
    twitter: string;
    linkedin: string;
    occupation: string;
    spoken_languages: string;
    address: string;
    biography: string;
    latitude: number;
    longitude: number;
    user_id: number;
    hierarchy_score: number;
    creation_date: Date;
    last_updated_date: Date;
}